Had determinism issues with stats.
Server and client (both running on the same machine) can apparently disagree on
the value of ceiling((X^1.8)/20.0), or maybe when calculating an average. I've
just encountered a case where the client said the character had 108 movement
points, whereas the server said they had only 107. As a result, the client
showed a path to the player that the server would not allow.

To fix this, the server's values are sent to the client. This is only
required for fields with values that affect how the client behaves,
i.e., max health and movement points.