Adds a way to do some additive CSS.
As baffling as it may appear, Cascading Style Sheets cannot actually do
cascading property values through multiple classes (you can only
override the previous value, not see it). You'd think there would be a
"current()" function which returns the not-yet-evaluated,
about-to-be-overridden current value of the attribute, but no, there is
not. There is no way to have any equivalent either, and since it's been
nearly a decade they've seen propositions to address this deficiency
without actually adding any, I wouldn't expect things to change within
the next decade either. Doesn't help that they seem to consider the
issue to be intertwined with the animation attributes, whereas it ought
to be a generic thing.